+++ src/components/QueryEditor/EditorRows.tsx
@@ -2,9 +2,11 @@ import React from 'react';
 
 import { Stack } from './Stack';
 
-interface EditorRowsProps {}
+interface EditorRowsProps {
+  children?: React.ReactNode;
+}
 
-export const EditorRows: React.FC<EditorRowsProps> = ({ children }) => {
+export const EditorRows = ({ children }: EditorRowsProps) => {
   return (
     <Stack gap={0.5} direction="column">
       {children}