+++ src/components/QueryEditor/types.ts
@@ -1,4 +1,4 @@
-import { JsonTree } from "react-awesome-query-builder";
+import { JsonTree } from 'react-awesome-query-builder';
 
 import {
   DataFrame,
@@ -7,20 +7,17 @@ import {
   SelectableValue,
   TimeRange,
   toOption as toOptionFromData,
-} from "@grafana/data";
+} from '@grafana/data';
 
-import { DataQuery } from "@grafana/schema";
+import { DataQuery } from '@grafana/schema';
 
-import { QueryWithDefaults } from "./defaults";
+import { QueryWithDefaults } from './defaults';
 import {
   QueryEditorFunctionExpression,
   QueryEditorGroupByExpression,
   QueryEditorPropertyExpression,
-} from "./expressions";
-import {
-  CompletionItemKind,
-  LanguageCompletionProvider,
-} from "@grafana/experimental";
+} from './expressions';
+import { CompletionItemKind, LanguageCompletionProvider } from '@grafana/experimental';
 
 export interface SqlQueryForInterpolation {
   dataset?: string;
@@ -59,8 +56,8 @@ export enum QueryFormat {
 }
 
 export enum EditorMode {
-  Builder = "builder",
-  Code = "code",
+  Builder = 'builder',
+  Code = 'code',
 }
 
 export interface SQLQuery extends DataQuery {
@@ -88,7 +85,7 @@ export interface SQLExpression {
   filters?: SQLFilters;
   groupBy?: QueryEditorGroupByExpression[];
   orderBy?: QueryEditorPropertyExpression;
-  orderByDirection?: "ASC" | "DESC";
+  orderByDirection?: 'ASC' | 'DESC';
   limit?: number;
   offset?: number;
 }
@@ -114,8 +111,8 @@ export interface QueryRowFilter {
 }
 
 export const QUERY_FORMAT_OPTIONS = [
-  { label: "Time series", value: QueryFormat.Timeseries },
-  { label: "Table", value: QueryFormat.Table },
+  { label: 'Time series', value: QueryFormat.Timeseries },
+  { label: 'Table', value: QueryFormat.Table },
 ];
 
 const backWardToOption = (value: string) => ({ label: value, value });
@@ -129,13 +126,7 @@ export interface ResourceSelectorProps {
 }
 // React Awesome Query builder field types.
 // These are responsible for rendering the correct UI for the field.
-export type RAQBFieldTypes =
-  | "text"
-  | "number"
-  | "boolean"
-  | "datetime"
-  | "date"
-  | "time";
+export type RAQBFieldTypes = 'text' | 'number' | 'boolean' | 'datetime' | 'date' | 'time';
 
 export interface SQLSelectableValue extends SelectableValue {
   type?: string;
@@ -153,19 +144,14 @@ export interface DB {
   datasets: () => Promise<string[]>;
   tables: (dataset?: string) => Promise<string[]>;
   fields: (query: SQLQuery, order?: boolean) => Promise<SQLSelectableValue[]>;
-  validateQuery: (
-    query: SQLQuery,
-    range?: TimeRange
-  ) => Promise<ValidationResults>;
+  validateQuery: (query: SQLQuery, range?: TimeRange) => Promise<ValidationResults>;
   dsID: () => number;
   dispose?: (dsID?: string) => void;
-  lookup: (
-    path?: string
-  ) => Promise<Array<{ name: string; completion: string }>>;
+  lookup: (path?: string) => Promise<Array<{ name: string; completion: string }>>;
   getSqlCompletionProvider: () => LanguageCompletionProvider;
   toRawSql?: (query: SQLQuery) => string;
   functions: () => Promise<Aggregate[]>;
-  labels?: Map<"dataset", string>;
+  labels?: Map<'dataset', string>;
   disableDatasets?: boolean;
 }