+++ src/components/QueryEditor/visual-query-builder/VisualEditor.tsx
@@ -13,10 +13,10 @@ import { SQLQuery } from '../types';
 import { TimeRange } from '@grafana/data';
 import { EditorRow } from './EditorRow';
 import { EditorField } from './EditorField';
-import { EditorRows} from './EditorRows';
+import { EditorRows } from './EditorRows';
 
 interface VisualEditorProps extends QueryEditorProps {
-  query: SQLQuery
+  query: SQLQuery;
   db: DB;
   queryRowFilter: QueryRowFilter;
   onChange: (q: SQLQuery) => void;
@@ -24,14 +24,7 @@ interface VisualEditorProps extends QueryEditorProps {
   range?: TimeRange;
 }
 
-export const VisualEditor: React.FC<VisualEditorProps> = ({
-  query,
-  db,
-  queryRowFilter,
-  onChange,
-  onValidate,
-  range,
-}) => {
+export const VisualEditor = ({ query, db, queryRowFilter, onChange, onValidate, range }: VisualEditorProps) => {
   const state = useAsync(async () => {
     const fields = await db.fields(query);
     return fields;