+++ src/unreleasedComponents/index.ts
@@ -1,3 +1,0 @@
-// noop for when there are no components here
-export { SecretInput } from "./SecretInput";
-export { SecretTextArea } from "./SecretTextarea";